Add `InputStream` versions of `TextExtraction` methods
Added `InputStream` versions of the `TextExtraction.uploadAndStartExtraction`
and `TextExtraction.uploadExtractAndFetch` methods. These versions provide
more flexibility to clients regarding where to store the file before uploading.

For instance, if the file is already stored in memory, the client can now
pass a `ByteArrayInputStream` view of the file's bytes for uploading,
rather than requiring a filesystem handle.